packages/cosmos-gradio/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cosmos-gradio"
version = "0.1.0"
description = "Gradio interface for Cosmos models"
license = {text = "Apache-2.0"}
authors = [
    {name = "NVIDIA Corporation"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "gradio>=4.0.0",
    "torch>=1.12.0",
    "loguru>=0.6.0",
    "pydantic-settings>=2.12.0",
    "pyparsing>=3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "mypy>=0.950",
]

[project.readme]
content-type = "text/markdown"
text = '''
# Cosmos-Gradio
'''

# UV settings must be present to avoid inheritance from the root pyproject.toml.
[tool.uv]

[tool.hatch.build.targets.sdist]
packages = [
  "cosmos_gradio",
]

[tool.hatch.build.targets.wheel]
packages = [
  "cosmos_gradio",
]
exclude = [
  "*_test.py",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
